Create between-two-cities.md
I am in the process of writing up summarised versions of many games - the idea being for use as refreshers or teaching aids. I understand this is not exactly what your site is doing, but take a look and maybe there is a place for it - differently tagged, or just as it is. If not, feel free to ignore.

Contents and tags are stubbed for now.

This game is Between Two Cities: https://boardgamegeek.com/boardgame/168435/between-two-cities